e2etest: enhance matrix comparison
Here, we are discontinuing the use of reflect.DeepEqual because
it may result in false negatives when the matrices have differences
we can ignore, such as order or service names.

Instead, we have implemented the test to compare the matrices using
the Diff function, and in the event of a failure, we print the difference.

Signed-off-by: Lior Noy <lnoy@redhat.com>