ss: Refactor ToComDetails for K8S Compatibility
We modify the ToComDetails function to make the 'rpc.statd'
process optional. Instead of creating ComDetails with two
service names, we extract the first item, which is the primary one.
This change is necessary because some 'ss' output entries return
two names separated by a comma, and this is an invalid name
for a Kubernetes object.

Signed-off-by: Lior Noy <lnoy@redhat.com>