Add new "nftables" module
The new 'nftables' module is responsible for creating
the nftables rules that block all ports except for the
ones in the communication matrix.

The function for this task is 'GetRulesFromCommDetails.'
It only returns an nftable file (or an error), and the
caller is responsible for executing the update on the node.

Signed-off-by: Lior Noy <lnoy@redhat.com>